requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
pycodestyle==2.5.0
pycparser==2.20
python-dotenv==0.12.0
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.4
